feat: refactor cjsToEsm.cjs file
